core: Moved bond handler execution to CellStructure
