+++ src/core/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(EspressoCore_SRC
     accumulators.cpp
+    bond_error.cpp
     cells.cpp
     collision.cpp
     comfixed_global.cpp
