+++ src/core/forces_inline.hpp
@@ -523,6 +523,6 @@ inline bool add_bonded_force(Particle &p1, int bond_id,
  *  @param p particle for which to calculate forces
  */
 inline void add_single_particle_force(Particle &p) {
-  execute_bond_handler(cell_structure, p, add_bonded_force);
+  cell_structure.execute_bond_handler(p, add_bonded_force);
 }
 #endif
