+++ src/core/object-in-fluid/oif_global_forces.cpp
@@ -65,33 +65,29 @@ void calc_oif_global(double *area_volume, int molType, CellStructure &cs) {
     if (p.p.mol_id != molType)
       continue;
 
-    execute_bond_handler(
-        cs, p,
-        [&partArea, &VOL_partVol](Particle &p1, int bond_id,
-                                  Utils::Span<Particle *> partners) {
-          auto const &iaparams = bonded_ia_params[bond_id];
+    cs.execute_bond_handler(p, [&partArea, &VOL_partVol](
+                                   Particle &p1, int bond_id,
+                                   Utils::Span<Particle *> partners) {
+      auto const &iaparams = bonded_ia_params[bond_id];
 
-          if (iaparams.type == BONDED_IA_OIF_GLOBAL_FORCES) {
-            // remaining neighbors fetched
-            auto const p11 =
-                unfolded_position(p1.r.p, p1.l.i, box_geo.length());
-            auto const p22 =
-                p11 + get_mi_vector(partners[0]->r.p, p11, box_geo);
-            auto const p33 =
-                p11 + get_mi_vector(partners[1]->r.p, p11, box_geo);
+      if (iaparams.type == BONDED_IA_OIF_GLOBAL_FORCES) {
+        // remaining neighbors fetched
+        auto const p11 = unfolded_position(p1.r.p, p1.l.i, box_geo.length());
+        auto const p22 = p11 + get_mi_vector(partners[0]->r.p, p11, box_geo);
+        auto const p33 = p11 + get_mi_vector(partners[1]->r.p, p11, box_geo);
 
-            // unfolded positions correct
-            auto const VOL_A = area_triangle(p11, p22, p33);
-            partArea += VOL_A;
+        // unfolded positions correct
+        auto const VOL_A = area_triangle(p11, p22, p33);
+        partArea += VOL_A;
 
-            auto const VOL_norm = get_n_triangle(p11, p22, p33);
-            auto const VOL_dn = VOL_norm.norm();
-            auto const VOL_hz = 1.0 / 3.0 * (p11[2] + p22[2] + p33[2]);
-            VOL_partVol += VOL_A * -1 * VOL_norm[2] / VOL_dn * VOL_hz;
-          }
+        auto const VOL_norm = get_n_triangle(p11, p22, p33);
+        auto const VOL_dn = VOL_norm.norm();
+        auto const VOL_hz = 1.0 / 3.0 * (p11[2] + p22[2] + p33[2]);
+        VOL_partVol += VOL_A * -1 * VOL_norm[2] / VOL_dn * VOL_hz;
+      }
 
-          return false;
-        });
+      return false;
+    });
   }
 
   part_area_volume[0] = partArea;
@@ -111,56 +107,50 @@ void add_oif_global_forces(double const *area_volume, int molType,
     if (p.p.mol_id != molType)
       continue;
 
-    execute_bond_handler(
-        cs, p,
-        [area, VOL_volume](Particle &p1, int bond_id,
-                           Utils::Span<Particle *> partners) {
-          auto const &iaparams = bonded_ia_params[bond_id];
+    cs.execute_bond_handler(p, [area,
+                                VOL_volume](Particle &p1, int bond_id,
+                                            Utils::Span<Particle *> partners) {
+      auto const &iaparams = bonded_ia_params[bond_id];
 
-          if (iaparams.type == BONDED_IA_OIF_GLOBAL_FORCES) {
-            auto const p11 =
-                unfolded_position(p1.r.p, p1.l.i, box_geo.length());
-            auto const p22 =
-                p11 + get_mi_vector(partners[0]->r.p, p11, box_geo);
-            auto const p33 =
-                p11 + get_mi_vector(partners[1]->r.p, p11, box_geo);
+      if (iaparams.type == BONDED_IA_OIF_GLOBAL_FORCES) {
+        auto const p11 = unfolded_position(p1.r.p, p1.l.i, box_geo.length());
+        auto const p22 = p11 + get_mi_vector(partners[0]->r.p, p11, box_geo);
+        auto const p33 = p11 + get_mi_vector(partners[1]->r.p, p11, box_geo);
 
-            // unfolded positions correct
-            // starting code from volume force
-            auto const VOL_norm = get_n_triangle(p11, p22, p33).normalize();
-            auto const VOL_A = area_triangle(p11, p22, p33);
-            auto const VOL_vv = (VOL_volume - iaparams.p.oif_global_forces.V0) /
-                                iaparams.p.oif_global_forces.V0;
+        // unfolded positions correct
+        // starting code from volume force
+        auto const VOL_norm = get_n_triangle(p11, p22, p33).normalize();
+        auto const VOL_A = area_triangle(p11, p22, p33);
+        auto const VOL_vv = (VOL_volume - iaparams.p.oif_global_forces.V0) /
+                            iaparams.p.oif_global_forces.V0;
 
-            auto const VOL_force = (1.0 / 3.0) *
-                                   iaparams.p.oif_global_forces.kv * VOL_vv *
-                                   VOL_A * VOL_norm;
+        auto const VOL_force = (1.0 / 3.0) * iaparams.p.oif_global_forces.kv *
+                               VOL_vv * VOL_A * VOL_norm;
 
-            auto const h = (1. / 3.) * (p11 + p22 + p33);
+        auto const h = (1. / 3.) * (p11 + p22 + p33);
 
-            auto const deltaA = (area - iaparams.p.oif_global_forces.A0_g) /
-                                iaparams.p.oif_global_forces.A0_g;
+        auto const deltaA = (area - iaparams.p.oif_global_forces.A0_g) /
+                            iaparams.p.oif_global_forces.A0_g;
 
-            auto const m1 = h - p11;
-            auto const m2 = h - p22;
-            auto const m3 = h - p33;
+        auto const m1 = h - p11;
+        auto const m2 = h - p22;
+        auto const m3 = h - p33;
 
-            auto const m1_length = m1.norm();
-            auto const m2_length = m2.norm();
-            auto const m3_length = m3.norm();
+        auto const m1_length = m1.norm();
+        auto const m2_length = m2.norm();
+        auto const m3_length = m3.norm();
 
-            auto const fac = iaparams.p.oif_global_forces.ka_g * VOL_A *
-                             deltaA /
-                             (m1_length * m1_length + m2_length * m2_length +
-                              m3_length * m3_length);
+        auto const fac = iaparams.p.oif_global_forces.ka_g * VOL_A * deltaA /
+                         (m1_length * m1_length + m2_length * m2_length +
+                          m3_length * m3_length);
 
-            p1.f.f += fac * m1 + VOL_force;
-            partners[0]->f.f += fac * m2 + VOL_force;
-            partners[1]->f.f += fac * m3 + VOL_force;
-          }
+        p1.f.f += fac * m1 + VOL_force;
+        partners[0]->f.f += fac * m2 + VOL_force;
+        partners[1]->f.f += fac * m3 + VOL_force;
+      }
 
-          return false;
-        });
+      return false;
+    });
   }
 }
